Cannot upload file upper than 50MB

+++ main.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from model.animeforce import Enforcer
+from src.model.animeforce import Enforcer
+from src.model.SqlLiteDB import sqliteDriver
+from src.config import TOKEN_API, TIME_INTERVAL
 import requests
-from model.SqlLiteDB import sqliteDriver
 from telegram.ext import Updater, CommandHandler
 import logging
 import time
-from src.config import TOKEN_API
-import urllib.request
+import os
 
 folder_path={
     'video': './src/video'
@@ -23,43 +23,33 @@ def delete_file(filename):
         print(e)
 
 def send_video_telegram(context, update, episode, filename):
+    print('Sending video to Telegram...')
     context.bot.send_video(
         chat_id=update.message.chat.id, 
         video=open(filename, 'rb'),
-        caption=f'#{episode.name} | Episode: {episode.number}'
-        #supports_streaming=True
+        caption=f'#{episode.name} | Episode: {episode.number}',
+        timeout=30000,
+        supports_streaming=True
     )
 
-def downloadVideo(context, update, episode):
+def download_video(context, update, episode):
     if not episode.link is None:
         url = episode.download_link()
         name = '-'.join(episode.name.split(' '))
         filename = f'{folder_path["video"]}/{name}.mp4'
-        print(url)
-        urllib.request.urlretrieve(url, filename)
+        if not os.path.isfile(filename):
+            chunk_size = 256
+            r = requests.get(url, stream=True)
+            with open(filename, "wb") as f:
+                print('Downloading Video...')
+                for chunk in r.iter_content(chunk_size=chunk_size):
+                    f.write(chunk)
         send_video_telegram(context, update, episode, filename)
         db.add_record(episode)
         delete_file(filename)
 
-def download_video(context, update, episode):
-    print(episode.link)
-    if not episode.link is None:
-        url = episode.download_link()
-        name = '-'.join(episode.name.split(' '))
-        filename = f'{folder_path["video"]}/{name}.mp4'
-        chunk_size = 256
-        r = requests.get(url, stream=True)
-        
-        with open(filename, "wb") as f:
-            print('Downloading Video...')
-            for chunk in r.iter_content(chunk_size=chunk_size):
-                f.write(chunk)
-            print('Sending video to Telegram...')
-            send_video_telegram(context, update, episode, filename)
-            db.add_record(episode)
-            delete_file(filename)
-
 def start(update, context):
+    print('Start main')
     enforcer = Enforcer()
     while True:
         anime_list = enforcer.last_episode_list()
@@ -67,23 +57,24 @@ def start(update, context):
             if anime.episode < 30:
                 last_episode = anime._get_last_episode()
                 if db.find_record(last_episode) is None:
+                    print(f'Downloading: {last_episode.name}')
                     download_video(context, update, last_episode)
-        time.sleep(3600)
-        
+        time.sleep(TIME_INTERVAL)
 
 def animelist(update, context):
     rows = db.get_animeList()
     msg = 'Anime List: \n'
     print(rows)
-    if not len(rows) == 0:
+    if rows:
         for r in rows:
             msg += f'#{r}\n'
-        update.message.reply_text(msg)
     else:
-        update.message.reply_text('Nessun anime trovato')
+        msg = 'Nessun anime trovato'
+    update.message.reply_text(msg)
 
 def main():
-    updater = Updater(token=TOKEN_API, use_context=True)
+    updater = Updater(token=TOKEN_API, use_context=True, request_kwargs={'read_timeout': 50, 'connect_timeout': 50})
+    print('Starting telegram bot...')
     dp = updater.dispatcher
     logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
                      level=logging.INFO)
@@ -91,6 +82,7 @@ def main():
     dp.add_handler(CommandHandler("animelist", animelist, pass_args=False))
     # Start the Bot
     updater.start_polling()
+    print('Telegram bot start')
     #updater.start_webhook(listen='127.0.0.1', port=5002, url_path=TOKEN_API)
     updater.idle()
